// week-04/starter/src/types.ts
export interface Paciente {
  id: number;
  nombre: string;
  tipoTratamiento: "hemodialisis" | "dialisis_peritoneal" | "hemodiafiltracion";
  turno: "mañana" | "tarde" | "noche";
  costoSesion: number;
  diasPorSemana: number;
  activo: boolean;
  createdAt: string;
}

export interface RespuestaLista<T> {
  data: T[];
  total: number;
  page: number;
  limit: number;
}

export interface RespuestaUnica<T> {
  data: T;
}